+++ networks/movement/movement-config/src/releases/mod.rs
@@ -1,0 +1,1 @@
+pub mod biarritz_rc1;
