+++ protocol-units/settlement/mcr/config/Cargo.toml
@@ -15,8 +15,8 @@ serde = { workspace = true , features = ["derive"] }
 alloy = { workspace = true }
 godfig = { workspace = true }
 anyhow = { workspace = true }
-
 movement-signer-loader = { workspace = true }
+movement-signer = { workspace = true}
 
 [lints]
 workspace = true
